Use a long long to store debug type flags. Fix a bug when it is impossible to log debug data in i386.

+++ inc/com/centreon/engine/configuration/state.hh
@@ -184,8 +184,8 @@ namespace               configuration {
     void                date_format(date_type value);
     std::string const&  debug_file() const throw ();
     void                debug_file(std::string const& value);
-    unsigned long       debug_level() const throw ();
-    void                debug_level(unsigned long value);
+    unsigned long long  debug_level() const throw ();
+    void                debug_level(unsigned long long value);
     unsigned int        debug_verbosity() const throw ();
     void                debug_verbosity(unsigned int value);
     bool                enable_environment_macros() const throw ();
@@ -535,7 +535,7 @@ namespace               configuration {
     set_contact         _contacts;
     date_type           _date_format;
     std::string         _debug_file;
-    unsigned long       _debug_level;
+    unsigned long long  _debug_level;
     unsigned int        _debug_verbosity;
     bool                _enable_environment_macros;
     bool                _enable_event_handlers;
